feat: add 4 models for messanger with migration

+++ WeMovieSync/WeMovieSync.Application/Interfaces/IUserRepository.cs
@@ -8,7 +8,7 @@ namespace WeMovieSync.Application.Interfaces
         Task<bool> EmailExistsAsync(string email);
         Task AddUserAsync(User user);
         Task<User?> GetByEmailAsync(string email);
-        Task<User?> GetByIdAsync(int id);
+        Task<User?> GetByIdAsync(long id);
         Task AddRefreshTokenAsync(RefreshToken token);
         Task<RefreshToken?> GetRefreshTokenAsync(string token);
         Task UpdateRefreshTokenAsync(RefreshToken token);
